Minor refactor to split off the uTox code from the Toxcore code
