+++ src/utox.h
@@ -1,0 +1,8 @@
+#ifndef UTOX_H
+#define UTOX_H
+
+#include <inttypes.h>
+
+void utox_message(uint8_t tox_message_id, uint16_t param1, uint16_t param2, void *data);
+
+#endif
